backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
pytesseract
google-generativeai
reportlab
supabase
python-dotenv
Pillow
pdf2image
